python_analytics/requirements.txt
pandas
sqlalchemy
psycopg2-binary
plotly
python-dotenv
